+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 
+- bugfix
+
 ## 6.1.0-IJ2022.2 - 2023-05-15
 
-- highlighting of functions and variables  [#389](https://github.com/esteinberg/plantuml4idea/issues/389)
+- highlighting of functions and variables #389
 - PlantUML library upgrade to v1.2023.7
 
 ## 6.0.0-IJ2022.2 - 2023-04-27