+++ CHANGELOG.md
@@ -2,5 +2,11 @@
 
 ## [Unreleased]
 
-- Rendering exception fix
+## [7.0.0-IJ2023.2] - 2023-08-25
 
+- SVG rendering migrated to EchoSVG from Batik
+- PlantUML library upgrade to v1.2023.10
+
+[Unreleased]: https://github.com/esteinberg/plantuml4idea/compare/v7.0.0-IJ2023.2...HEAD
+
+[7.0.0-IJ2023.2]: https://github.com/esteinberg/plantuml4idea/commits/v7.0.0-IJ2023.2